Started work on Distance Tool
